+++ controllers/KeyclockController/ScopeController.js
@@ -5,7 +5,13 @@ const {
     updateScope,
 } = require("../../middlewares/keyclock/Scope") ;
 const { error } = require( "../../utils/errorMessages");
-const { getTokenFromRequestHeader } = require("../../middlewares/commonFunctions")
+const { getTokenFromRequestHeader } = require("../../middlewares/commonFunctions");
+
+const {
+    checkValidityToCreateScope,
+    checkValidityToDeleteScope,
+    checkValidityToUpdateScope,
+} = require("../../middlewares/validators/ScopeValidators")
 
 exports.createScope = async (req, res) => {
     let token = getTokenFromRequestHeader(req,res);
@@ -13,12 +19,14 @@ exports.createScope = async (req, res) => {
         name,
         displayName,
     } = req.body;
-    const createScopeResponse = await createScope(token,name,displayName);
-    if(JSON.parse(createScopeResponse).error){
-        res.status(200).json(error.SCOPE_CREATE);
-    }else{
-        res.send({code: 1,message: "Scope created successfully" });
-    }      
+    if(checkValidityToCreateScope(req,res)){
+        const createScopeResponse = await createScope(token,name,displayName);
+        if(JSON.parse(createScopeResponse).error){
+            res.status(200).json(error.SCOPE_CREATE);
+        }else{
+            res.send({code: 1,message: "Scope created successfully" });
+        }
+    }    
 };
 
 exports.getAllScopes = async (req, res) => {
@@ -39,6 +47,7 @@ exports.getAllScopes = async (req, res) => {
 exports.deleteScopes = async (req, res) => {
     const { scopeId } = req.body;
     let token = getTokenFromRequestHeader(req,res);
+    if(checkValidityToDeleteScope(req,res)){
         const deleteScopeResponse = await deleteScope(token, scopeId);
         console.log(deleteScopeResponse);
         if(deleteScopeResponse.length !== 0 ){
@@ -49,11 +58,13 @@ exports.deleteScopes = async (req, res) => {
                 message: "Scope Deleted successfully"
             });
         }
+    }
 };
 
 exports.updateScopes = async (req, res) => {
     const { scopeId, name, displayName} = req.body;
     let token = getTokenFromRequestHeader(req,res);
+    if(checkValidityToUpdateScope(req,res)){
         const updateScopeResponse = await updateScope(token, scopeId, name, displayName);
         console.log(updateScopeResponse);
         if(updateScopeResponse.length !== 0 ){
@@ -64,4 +75,5 @@ exports.updateScopes = async (req, res) => {
                 message: "Scope Updated successfully"
             });
         }
+    }
 };
